+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(components-squares): Squares component added
 * 🎉 feat(login): fully responsive login page styles added
 
 .. 1.0.0 (2022-06-22)